File systems: add AddFS and change attach arguments
AddFS is how a given server names what is servers. Trying to add
a file system when one exists is fatal.

Attach arguments are now the name and an array of strings. In the next rev,
if a server implements Attach, it will be called, but it does not have to implement
it.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>